requirements.txt
Flask==1.0.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.9
matplotlib==3.3.2
numpy==1.16.2
tensorflow=2.0.0
picamera
opencv-python
pyserial